[MOV] test_related_translation: from test_orm to test_translate
