+++ odoo/addons/test_orm/security/ir.access.csv
@@ -111,10 +111,6 @@ access_test_orm_prefetch_line,access_test_orm_prefetch_line,test_orm.prefetch.li
 access_test_orm_modified,access_test_orm_modified,test_orm.modified,base.group_user,crud,
 access_test_orm_modified_line,access_test_orm_modified_line,test_orm.modified.line,base.group_user,crud,
 access_test_orm_modified_line_positive,access_test_orm_modified_line_positive,test_orm.modified.line.positive,base.group_user,crud,
-access_test_orm_related_translation_1,access_test_orm_related_translation_1,test_orm.related_translation_1,base.group_system,crud,
-access_test_orm_related_translation_2,access_test_orm_related_translation_2,test_orm.related_translation_2,base.group_system,crud,
-access_test_orm_related_translation_3,access_test_orm_related_translation_3,test_orm.related_translation_3,base.group_system,crud,
-access_test_orm_related_translation_4,access_test_orm_related_translation_4,test_orm.related_translation_4,base.group_system,crud,
 access_test_orm_indexed_translation,access_test_orm_indexed_translation,test_orm.indexed_translation,base.group_system,crud,
 access_test_orm_empty_char,access_test_orm_empty_char,test_orm.empty_char,base.group_system,crud,
 access_test_orm_empty_int,access_test_orm_empty_int,test_orm.empty_int,base.group_system,crud,
