+++ odoo/addons/test_orm/tests/__init__.py
@@ -35,7 +35,6 @@ from . import (
     test_read_grouping_sets,
     test_registry,
     test_registry_signaling,
-    test_related_translation,
     test_schema,
     test_search,
     test_search_order,
