+++ odoo/addons/test_translation/models/__init__.py
@@ -1,1 +1,2 @@
 from . import models
+from . import test_related_translation
