+++ odoo/addons/test_translation/security/ir.access.csv
@@ -1,2 +1,6 @@
 id,name,model_id,group_id/id,operation,domain
 access_test_translation_constraint,access.test.translation.constraint,test.translation.constraint,base.group_user,crud,
+access_test_translation_related_translation_1,access_test_translation_related_translation_1,test_translation.related_translation_1,base.group_system,crud,
+access_test_translation_related_translation_2,access_test_translation_related_translation_2,test_translation.related_translation_2,base.group_system,crud,
+access_test_translation_related_translation_3,access_test_translation_related_translation_3,test_translation.related_translation_3,base.group_system,crud,
+access_test_translation_related_translation_4,access_test_translation_related_translation_4,test_translation.related_translation_4,base.group_system,crud,
