+++ odoo/addons/test_translation/tests/__init__.py
@@ -1,3 +1,4 @@
 from . import test_export_wizard
+from . import test_related_translation
 from . import test_term_count
 from . import test_ui
